% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_similarity.R
\name{batches_analysis}
\alias{batches_analysis}
\title{batches_analysis}
\usage{
batches_analysis(all_groups, n_cores, final_df, all_plots_names,
  paths_all_plots, thr_score, progress_bar = T)
}
\arguments{
\item{all_groups}{All batches names to analyze}

\item{n_cores}{Number to use}

\item{final_df}{Dataframe of features reduced}

\item{all_plots_names}{All plots names under analysis}

\item{paths_all_plots}{Path of all plots under analysis}

\item{thr_score}{Threshold for filtering the similarity scores}

\item{progress_bar}{If False,disable progress bar.}
}
\value{
List of results batch analysis
}
\description{
function to analyze each batch. Internal function.
}
\keyword{Internal}
