% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_visnet_conversion.R
\name{contract_data}
\alias{contract_data}
\title{contract_data}
\usage{
contract_data(data)
}
\arguments{
\item{data}{List of the full VisNetwork dataframe containing the edges and nodes information}
}
\value{
A list of two VisNetwork dataframes containing the contracted information of nodes and edges
}
\description{
function to contract the visnetdata object (containing nodes and edges info). Internal function.
}
\keyword{Internal}
