% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_exporting.R
\name{exports_filtered_plots}
\alias{exports_filtered_plots}
\title{exports_filtered_plots}
\usage{
exports_filtered_plots(visnetdata, df_features, path_expr_data,
  path_output = NULL, n_cores = 1, filter_thr = 2.1,
  plot_dendrogram = F, seed_n = 40)
}
\arguments{
\item{visnetdata}{list of dataframes containing the edges and nodes information in VisNetwork format}

\item{df_features}{A dataframe containing the original features set (before the tsne reduction) of all input files}

\item{path_expr_data}{Path to directory containing the expression data of the files analyzed}

\item{path_output}{Path to the directory where to export the plots for each group
If path_output=NULL, no plots are generated. Default to NULL}

\item{n_cores}{Number of cores to use. Default to 1}

\item{filter_thr}{float number indicating the cut height for the dendrogram generated
by the hierarchical clustering. Default to 2.1}

\item{plot_dendrogram}{If True, the plot of the dendrogram is generated (It may fail for some groups). Default to False.}

\item{seed_n}{Set seed number. Default to 40.}
}
\value{
A list of the selected samples names for each group
}
\description{
function to export only the automatically selected plots for each group
}
\examples{
\donttest{exports_filtered_plots(visnetdata=visnetdata,df_features=df_features,
path_expr_data="path to input directory",path_output=NULL,n_cores=1,filter_thr=2.1,plot_dendrogram=F)}
}
