% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_exporting.R
\name{exports_plots}
\alias{exports_plots}
\title{exports_plots}
\usage{
exports_plots(visnetdata, path_expr_data, path_output = NULL,
  n_cores = 1, path_gates = NULL)
}
\arguments{
\item{visnetdata}{list of dataframes containing the edges and nodes information in VisNetwork format}

\item{path_expr_data}{Path to directory containing the expression data of the files analyzed}

\item{path_output}{Path to the directory where to export the plots for each group
If path_output=NULL, no plots are generated. Default to NULL}

\item{n_cores}{Number of cores to use. Default to 1}

\item{path_gates}{Path of gates data, default to NULL}
}
\value{
A list of the samples names for each group.
}
\description{
function to export all plots for each group
}
\examples{
\donttest{exports_plots(visnetdata=visnetdata,path_expr_data="path to input directory",path_output=NULL,n_cores=1)}
}
