% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim-package.R
\docType{package}
\name{flowSim-package}
\alias{flowSim}
\alias{flowSim-package}
\title{flowSim: visualize, identify and remove redudant information from a bivariate FCM dataset}
\description{
To learn more about using the flowSim package, please start by browsing the
vignettes:
}
\seealso{
Useful links:
\itemize{
  \item \url{https://davidrach.github.io/flowSim/}
  \item Report bugs at \url{https://github.com/semontante/flowSim/issues}
}

}
\author{
\strong{Maintainer}: Sebastiano Montante \email{semont93@gmail.com} (\href{https://orcid.org/0000-0002-6717-8706}{ORCID})

}
\keyword{internal}
