% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_visualization.R
\name{flowSim_plot}
\alias{flowSim_plot}
\title{flowSim_plot}
\usage{
flowSim_plot(df, plot_gate = F)
}
\arguments{
\item{df}{Dataframe containing the two markers expression data.}

\item{plot_gate}{show gates?,default to False}
}
\value{
Plot object
}
\description{
function to generate the scatter plot with colored density of the events. Internal function.
}
\keyword{Internal}
