% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_visualization.R
\name{flowSim_plot_dens}
\alias{flowSim_plot_dens}
\title{flowSim_plot_dens}
\usage{
flowSim_plot_dens(df)
}
\arguments{
\item{df}{Dataframe containing the two markers expression data.}
}
\value{
Plot object
}
\description{
function to generate the denity scatter plot with density on the sides. Internal function.
}
\keyword{Internal}
