% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_igraph_clustering.R
\name{gen_igraph_network}
\alias{gen_igraph_network}
\title{Generate igraph network}
\usage{
gen_igraph_network(similarity_output, method_clust = "louvain")
}
\arguments{
\item{similarity_output}{The distance matrix generated by the get_similarity_all_plots functions}

\item{method_clust}{clustering method: louvain, infomap. Default to louvain}
}
\value{
two objects
$igraph_network : igraph network object
$partition : vector of the clusters association
}
\description{
function to generate the igraph network from similiraty scores
}
\examples{
\donttest{gen_igraph_network(similarity_output=out,method_clust="louvain")}
}
