% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_utils.R
\name{get_csv}
\alias{get_csv}
\title{get_csv}
\usage{
get_csv(path, channels = NULL, markers = NULL, path.output)
}
\arguments{
\item{path}{path of the directory with fcs files to convert}

\item{channels}{Vector reporting the two channels to consider}

\item{markers}{Vector reporting the two markers to consider. Considered only if channels==NULL}

\item{path.output}{directory path to export the converted csv files}
}
\description{
function to convert fcs files in bivariate csv files
}
\examples{
\donttest{get_csv(path="path/to/directory",channels=c("FSC-A","FSC-H"),path.output="path/to/directory")}
}
