% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_visualization.R
\name{get_densRange}
\alias{get_densRange}
\title{get_densRange}
\usage{
get_densRange(x, y, thr, direction = FALSE)
}
\arguments{
\item{x}{x axis density coordinates}

\item{y}{y axis density coordinates}

\item{thr}{Treshold of expression where to start the density calculation}

\item{direction}{On the left (FALSE) or on the right (TRUE) of the threshold}
}
\value{
Coordinates density points for the x and y axis
}
\description{
function to get the correct coordinates (correct range) based on the density
coordinates x and y generated by the density() function. Internal function.
}
\keyword{Internal}
