% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_utils.R
\name{get_distance_loc_vs_test}
\alias{get_distance_loc_vs_test}
\title{get distance selected file vs another local file}
\usage{
get_distance_loc_vs_test(test_df, loc_df, show_plot = "none",
  nboot = 50)
}
\arguments{
\item{test_df}{Dataframe of selected file}

\item{loc_df}{Dataframe of another file of the input directory}

\item{show_plot}{show plot of the density comparison}

\item{nboot}{Number of simulations for the permuation identity test between the densities.}
}
\value{
A list of two float numbers
$pvalue_mark_1 : pvalue first marker comparison
$pvalue_mark_2 : pvalue second marker comparison
}
\description{
function to compare the similiraty between selected file and
another file of the input directory based on the density of marker expression values. Internal function.
}
\keyword{Internal}
