% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_igraph_clustering.R
\name{get_groups_based_on_clustering}
\alias{get_groups_based_on_clustering}
\title{Generate groups based on a clustering method}
\usage{
get_groups_based_on_clustering(igraph_network, method = "louvain")
}
\arguments{
\item{similarity_output}{Output generated by the get_similarity_all_plots or get_similarity_all_plots_v2 function}

\item{method_clust}{clustering method: louvain, infomap, default to louvain}
}
\value{
two objects
$igraph_network : igraph network object
$partition : vector of the clusters association
}
\description{
function to clusterize the samples using different methods. Internal function.
}
\keyword{Internal}
