% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_utils.R
\name{get_heterogeneity_score}
\alias{get_heterogeneity_score}
\title{Get heterogeneity score}
\usage{
get_heterogeneity_score(igraph_network, partition_vec, visnetdata)
}
\arguments{
\item{igraph_network}{Igraph network object}

\item{partition_vec}{Vectors of files and clusters associations}

\item{visnetdata}{list of dataframes containing the edges and nodes information in VisNetwork format}
}
\value{
A single float number indicating the final heterogeneity score
}
\description{
function to get the heterogeneity score of a dataset
}
\examples{
\donttest{gen_igraph_network(igraph_network=igraph_network,partition_vec=partition_vec,visnetdata=visnetdata)}
}
