% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_utils.R
\name{get_hull_all_gates}
\alias{get_hull_all_gates}
\title{get_hull_all_gates}
\usage{
get_hull_all_gates(gated_df)
}
\arguments{
\item{df_groups}{Dataframe containing the groups associations}
}
\value{
A list containing the convex hull for all classes
}
\description{
function to get the convex hull of all gates. Internal function
}
\keyword{Internal}
