% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_utils.R
\name{get_inds_files_selected}
\alias{get_inds_files_selected}
\title{get_inds_files_selected}
\usage{
get_inds_files_selected(path_dir, files_selected, return_path = F)
}
\arguments{
\item{path_dir}{path of the directory to explore}

\item{files_selected}{Vector containing the names of the files whose indices need to be found}

\item{return_path}{If True, return paths instead of indices. Default to False}
}
\value{
Vector containing the indices of the files
}
\description{
function to get indices for the paths of the files selected
}
\examples{
\donttest{get_inds_files_selected(path_dir="path/to/directory",files_selected=vector_names)}
}
