% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_similarity.R
\name{get_similarity_all_plots}
\alias{get_similarity_all_plots}
\title{Generate similarity scores all files V1 mode}
\usage{
get_similarity_all_plots(n_samples = NULL, path_dir = NULL,
  thr_score = 0.6, n_cores = 1, nboots = NULL, paths_plots = NULL)
}
\arguments{
\item{n_samples}{Indices of files paths extracted from list.files(path_dir). If NULL, all files in path_dir are analyzed. Default to NULL.}

\item{path_dir}{Path to tthe input directory}

\item{thr_score}{Threshold for filtering only the most significant similarity score, it ranges from 0 to 1. Default to 0.6.
Note: thr=0, keep all edges except edges with 0 weight. To keep all edges set thr="all".}

\item{n_cores}{Number of cores to use, default to 1}

\item{nboots}{Number of bootstrap simulations for the density comparison. If NULL, it is set based on number of files anayzed. Default to NULL}

\item{paths_plots}{Mandatory if paths_dir==NULL, it is possible to feed a list of the paths of the input files to analyze}
}
\value{
A list of two dataframes
$edges : A dataframe containing edges information of the network. The weight of each edge is the similarity score between the the two files/nodes
$nodes : A dataframe containing several nodes information (Node ID,cluster color etc...)
}
\description{
function to generate the similarity scores for all the files within the input directory. V1 mode (for small datasets only,N<500).
}
\examples{
\donttest{get_similarity_all_plots(path_dir="path to input directory",thr_score=0.9,n_cores=1)}
}
