% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_similarity.R
\name{get_similarity_all_plots_v2}
\alias{get_similarity_all_plots_v2}
\title{Generate similarity scores all files V2 mode}
\usage{
get_similarity_all_plots_v2(n_samples = NULL, path_dir,
  thr_score = 0.9, n_cores = 1, n_batches = NULL,
  show_tsne_plot = T, progress_bar = T, size_points_tsne = 5,
  nboots = 6)
}
\arguments{
\item{n_samples}{Indices of files paths extracted from list.files(path_dir). If NULL, all files in path_dir are analyzed. Default to NULL.}

\item{path_dir}{Path to input directory. Mandatory.}

\item{thr_score}{Treshold for filtering only the most significant similarity score, it ranges from 0 to 1. Default to 0.9}

\item{n_cores}{Number of cores to use, default to 1}

\item{n_batches}{Number of batches to use = number of parts in which the input datasets must be divided to
increase speed and memory efficiency. If number of files <= 2000 files, this integer is automatically estimated,
otherwise it needs to be defined. As the value increases, less memory and time is required but also more residual homogeneity is generated.}

\item{show_tsne_plot}{show scatter plot of tsne reduced dataset of the density features. Default to True.}

\item{progress_bar}{if False,disable progress bar. Default to True.}

\item{size_points_tsne}{Size of points in t-sne plot. Default to 5}

\item{nboots}{Number of bootstrap simulations for the  density identiy test. Default to 6}
}
\value{
A list of three dataframes
$edges : A dataframe containing edges information of the network. The weight of each edge is the similarity score between the the two files/nodes
$nodes : A dataframe containing several nodes information (Node ID,cluster color etc...)
$df_features: A dataframe containing the original features set (before the tsne reduction) of all input files
}
\description{
function to generate the similarity scores for all the files within the input directory. V2 mode (for big datasets)
}
\examples{
\donttest{get_similarity_all_plots_v2(n_samples=1:2000,path_dir="path to input directory",thr_score=0.9,n_cores=1,n_batches=NULL,show_tsne_plot=T)}
}
