% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_similarity.R
\name{get_similarity_selected_plot}
\alias{get_similarity_selected_plot}
\title{Get similarity score selected file}
\usage{
get_similarity_selected_plot(list_all_input_dfs, name_ref_plot, n_cores,
  names_analyzed_vec, n_boots)
}
\arguments{
\item{list_all_input_dfs}{List of the dataset of the input directory generated by the import_all_dfs function}

\item{name_ref_plot}{file to analyze}

\item{n_cores}{Number of cores to use}

\item{names_analyzed_vec}{Vector containing the names of the files already analyzed in a previous iteration}

\item{nboots}{Number of bootstrap simulations for the density comparison.}
}
\value{
Dataframe of the similarity scores for the selected file
}
\description{
function to get the similarity of one file vs the rest. Internal usage only.
}
\keyword{Internal}
