% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_utils.R
\name{get_size_based_on_groups}
\alias{get_size_based_on_groups}
\title{get_size_based_on_groups}
\usage{
get_size_based_on_groups(df_groups)
}
\arguments{
\item{df_groups}{Dataframe containing the groups associations}
}
\value{
A dataframe containing the association between groups,files and size
}
\description{
function to get size of groups based on groups. Internal function
}
\keyword{Internal}
