% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_importing.R
\name{import_all_dfs}
\alias{import_all_dfs}
\title{Import files to analyze}
\usage{
import_all_dfs(n_samples = NULL, paths_dir = NULL,
  paths_plots = NULL, n_cores = 1)
}
\arguments{
\item{n_samples}{Number of files to analyze. If NULL, all files are analyzed Default to NULL.}

\item{paths_dir}{directory path of the input files to analyze}

\item{paths_plots}{Mandatory if paths_dir==NULL, it is possible to feed a list of the paths of the input files to analyze}

\item{n_cores}{The number of cores to use,default to 1}
}
\value{
list of all input dfs
}
\description{
function to import all datasets in the input directory
}
\examples{
import_all_dfs(paths_dir="path_to_input_directory",paths_plots=NULL,n_cores=1)
}
