% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_importing.R
\name{import_df_features}
\alias{import_df_features}
\title{Import features set of all files}
\usage{
import_df_features(n_samples = NULL, paths_dir, n_cores = 1,
  progress_bar = T)
}
\arguments{
\item{n_samples}{number of files to analyze}

\item{paths_dir}{directory path of the input files to analyze}

\item{n_cores}{The number of cores to use, default to 1}

\item{progress_bar}{If False,progress bar is disabled. Default to True}
}
\value{
a dataframe of the features set (columns) for each file (row)
}
\description{
function to import the features set of all the files under analysis
}
\examples{
import_df_features(n_samples=1:5,paths_dir="path_to_input_directory",n_cores=1)
}
