% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_similarity.R
\name{myKmeans}
\alias{myKmeans}
\title{kmeans function}
\usage{
myKmeans(dist, k)
}
\arguments{
\item{dist}{dataframe of the tsne reduced features set between the files}

\item{k}{Number of centers}
}
\value{
k-means clustering output
}
\description{
function to perform k-means clustering on the tsne reduced dataset of density features. Internal function.
}
\keyword{Internal}
