% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_visualization.R
\name{show_plot}
\alias{show_plot}
\title{show_plot}
\usage{
show_plot(path_data, name_plot, type = "classic")
}
\arguments{
\item{path_data}{Path to the directory with all the files}

\item{name_plot}{Name of the file to show}

\item{type}{type of visualization: "classic" or "dens". Default to classic.
the "dens" version has the densities of the markers on the side of each axis.}
}
\value{
Plot object
}
\description{
function to visualize the selected plot using the flowSim_plot or flowSim_plot_dens function
}
\examples{
\donttest{show_plot(path_data="path to input directory",type="classic")}
}
