% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_visnet_conversion.R
\name{gen_visnetwork_data}
\alias{gen_visnetwork_data}
\title{gen_visnetwork_data}
\usage{
gen_visnetwork_data(
  igraph_network,
  contract_net = FALSE,
  single_nodes_black = TRUE
)
}
\arguments{
\item{igraph_network}{igraph network object}

\item{contract_net}{If TRUE, it generates a contracted version of the 
VisNetwork dataframes (to visualize large datasets). Default to FALSE}

\item{single_nodes_black}{Isolated files/nodes colored in black? Default to TRUE.}
}
\value{
If contracted_net=FALSE, a list of full VisNetwork dataframes is generated, if 
contracted_net=TRUE, a list of contracted VisNetwork dataframes is returned.
$nodes : dataframe containing the nodes information (full or contracted)
$edges : dataframe containing the edges information (full or contracted)
}
\description{
function to convert igraph network to a list of two VisNetwork dataframes containig 
the edges and nodes information needed to generate a VisNetwork object
}
\examples{
A <- 2+2

}
