% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_igraph_clustering.R
\name{get_groups_based_on_clustering}
\alias{get_groups_based_on_clustering}
\title{Generate groups based on a clustering method}
\usage{
get_groups_based_on_clustering(igraph_network, method = "louvain")
}
\arguments{
\item{igraph_network}{iGraph network object}

\item{method_clust}{clustering method: louvain, infomap. Default to louvain}
}
\value{
dataframe reporting groups association
}
\description{
function to clusterize the samples using different methods. Internal function.
}
\keyword{Internal}
