% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_similarity.R
\name{get_similarity_all_plots}
\alias{get_similarity_all_plots}
\title{Generate similarity scores all files first version}
\usage{
get_similarity_all_plots(n_samples = NULL, path_dir, thr_score = 0.9,
  n_cores = 1)
}
\arguments{
\item{n_samples}{Number of files to analyze. If NULL, all files are analyzed Default to NULL.}

\item{path_dir}{Path to tthe input directory}

\item{thr_score}{Threshold for filtering only the most significant similarity score, it ranges from 0 to 1. Default to 0.9}

\item{n_cores}{Number of cores to use, default to 1}
}
\value{
A list of two dataframes
$edges : A dataframe containing edges information of the network. The weight of each edge is the similarity score between the the two files/nodes
$nodes : A dataframe containing several nodes information (Node ID,cluster color etc...)
}
\description{
function to generate the similarity scores for all the files within the input directory. First version (for small datasets only,N<500).
}
\examples{
\donttest{get_similarity_all_plots(path_dir="path to input directory",thr_score=0.9,n_cores=1)}
}
