% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowSim_utils.R
\name{rename_files}
\alias{rename_files}
\title{rename_files}
\usage{
rename_files(path_dir, to_replace, string, to_replace_fixed = TRUE)
}
\arguments{
\item{path_dir}{path of the directory with files to rename}

\item{to_replace}{Pattern to replace. if to_replace_fixed=FALSE, regex expression can be used.}

\item{string}{string to replace the pattern selected}

\item{to_replace_fixed}{if False, to_replace can be also a regex expression. Default to True.}
}
\description{
Function to rename files within a directory
}
\examples{
A <- 2+2

}
